micro: extract composer context UI
